+++ src/components/base/index.ts
@@ -2,3 +2,7 @@ export { default as Button } from './Button/Button';
 export { default as ChoiceButton } from './ChoiceButton';
 export { default as Input } from './Input';
 export { default as Modal } from './Modal';
+export { default as CheckBox } from './CheckBox';
+export { default as MultiRangeSlider } from './MultiRangeSlider';
+export { default as ProgressBar } from './ProgressBar';
+export { default as Test } from './Test';
