--- src/components/domain/index.ts
@@ -1,1 +1,0 @@
-export { default as TypeOfMeeting } from './TypeOfMeeting';
