+++ src/components/domain/landing/index.ts
@@ -1,0 +1,2 @@
+export { default as LandingContainer } from './LandingContainer';
+export { default as LandingFooter } from './LandingFooter';
