+++ src/components/domain/survey/TypeOfMeeting.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { ChoiceButton } from '../base';
+import { ChoiceButton } from '../../base';
 import { ChoiceOptions } from '@/pages/TypeOfMeetingSurvey';
 import { SubTitle } from '@/lib/styles/styledComponents';
