+++ src/components/domain/survey/index.ts
@@ -1,0 +1,2 @@
+export { default as TypeOfMeeting } from './TypeOfMeeting';
+export { default as SurveyTemplate } from './SurveyTemplate';
