+++ src/pages/index.ts
@@ -1,0 +1,6 @@
+export { default as AuthMail } from './AuthMail';
+export { default as FoundPath } from './FoundPath';
+export { default as LandingPage } from './LandingPage';
+export { default as NotFound } from './NotFound';
+export { default as Survey } from './Survey';
+export { default as TypeOfMeetingSurvey } from './TypeOfMeetingSurvey';
